Rename folder because of typo
- singlePageFromSections -> singlePageFormSections
